more scala way to handle async image download
