+++ uploader/src/main/scala/net/hearthstats/CardUtils.scala
@@ -1,10 +1,16 @@
 package net.hearthstats
 
-import java.io.IOException
+import java.io.FileOutputStream
+import java.net.URL
+import java.nio.channels.Channels
+
+import scala.collection.JavaConversions.asScalaBuffer
+import scala.concurrent._
+import scala.util.Failure
+import scala.util.Success
+import scala.concurrent.duration._
+
 import net.hearthstats.log.Log
-import org.json.simple.JSONObject
-//remove if not needed
-import scala.collection.JavaConversions._
 
 object CardUtils {
   private lazy val api: API = new API
@@ -22,5 +28,27 @@ object CardUtils {
       cost = cost,
       name = json.get("name").toString,
       collectible = json.get("collectible").toString.toBoolean)).toMap
+
+  def downloadImages(cards: List[Card]) {
+    import ExecutionContext.Implicits.global
+    val futures = for (card <- cards) yield future {
+      val rbc = Channels.newChannel(new URL(card.url).openStream)
+      val file = card.localFile
+      if (file.length < 30000) {
+        val fos = new FileOutputStream(file)
+        fos.getChannel.transferFrom(rbc, 0, Long.MaxValue)
+        fos.close()
+        rbc.close()
+        Log.debug(card.name + " saved to cache folder")
+      } else
+        Log.debug(card.name + " already in cache, skipping")
+    }
+    val all = Future.sequence(futures)
+    all.onComplete {
+      case Success(_) => Log.info("all images downloaded successfully")
+      case Failure(e) => Log.warn("could not download an image", e)
+    }
+    Await.result(all, 10.seconds)
+  }
 
 }
