+++ uploader/src/main/scala/net/hearthstats/ui/ClickableDeckBox.scala
@@ -3,22 +3,19 @@ package net.hearthstats.ui
 import java.awt.Dimension
 import java.awt.event.MouseAdapter
 import java.awt.event.MouseEvent
-import java.io.FileOutputStream
-import java.net.URL
-import java.nio.channels.Channels
-import java.util.concurrent.Executors
-import java.util.concurrent.TimeUnit
+
 import scala.swing.Swing.onEDT
+
 import javax.swing.Box.createVerticalBox
 import javax.swing.BoxLayout
 import javax.swing.ImageIcon
 import javax.swing.JLabel
 import javax.swing.JPanel
 import net.hearthstats.Card
+import net.hearthstats.CardUtils
 import net.hearthstats.Deck
 import net.hearthstats.log.Log
 import net.hearthstats.logmonitor.CardDrawnObserver
-import net.hearthstats.ui.ClickableDeckBox.MouseHandler
 
 class ClickableDeckBox(deck: Deck) extends JPanel with CardDrawnObserver {
 
@@ -56,42 +53,14 @@ class ClickableDeckBox(deck: Deck) extends JPanel with CardDrawnObserver {
 object ClickableDeckBox {
 
   def makeBox(deck: Deck): ClickableDeckBox = {
-    downloadImages(deck)
+    CardUtils.downloadImages(deck.cards)
     new ClickableDeckBox(deck)
   }
+}
 
-  private def downloadImages(deck: Deck) {
-    val scheduledExecutorService = Executors.newScheduledThreadPool(30)
-    for (card <- deck.cards) {
-      scheduledExecutorService.submit(new Runnable {
+case class MouseHandler(card: Card, imageLabel: JLabel) extends MouseAdapter {
 
-        override def run {
-          try {
-            val rbc = Channels.newChannel(new URL(card.url).openStream)
-            val file = card.localFile
-            if (file.length < 30000) {
-              val fos = new FileOutputStream(file)
-              fos.getChannel.transferFrom(rbc, 0, Long.MaxValue)
-              fos.close()
-              rbc.close()
-              Log.debug(card.name + " saved to cache folder")
-            } else
-              Log.debug(card.name + " already in cache, skipping")
-          } catch {
-            case e: Exception => Log.error("Could not download image for " + card.name, e)
-          }
-        }
-      })
-    }
-    scheduledExecutorService.shutdown
-    scheduledExecutorService.awaitTermination(30, TimeUnit.SECONDS)
-    Log.info("all images downloaded successfully")
-  }
-
-  private class MouseHandler(var card: Card, var imageLabel: JLabel) extends MouseAdapter {
-
-    override def mouseEntered(e: MouseEvent) {
-      onEDT(imageLabel.setIcon(new ImageIcon(card.localURL)))
-    }
+  override def mouseEntered(e: MouseEvent) {
+    onEDT(imageLabel.setIcon(new ImageIcon(card.localURL)))
   }
 }
